chore: split functionality internally
